move common metrics-related funcs to internal package
