+++ modules/caddyhttp/metrics_test.go
@@ -5,7 +5,6 @@ import (
 	"errors"
 	"net/http"
 	"net/http/httptest"
-	"strings"
 	"testing"
 
 	"github.com/prometheus/client_golang/prometheus/testutil"
@@ -82,26 +81,4 @@ type middlewareHandlerFunc func(http.ResponseWriter, *http.Request, Handler) err
 
 func (f middlewareHandlerFunc) ServeHTTP(w http.ResponseWriter, r *http.Request, h Handler) error {
 	return f(w, r, h)
-}
-
-func TestSanitizeMethod(t *testing.T) {
-	tests := []struct {
-		method   string
-		expected string
-	}{
-		{method: "get", expected: "GET"},
-		{method: "POST", expected: "POST"},
-		{method: "OPTIONS", expected: "OPTIONS"},
-		{method: "connect", expected: "CONNECT"},
-		{method: "trace", expected: "TRACE"},
-		{method: "UNKNOWN", expected: "OTHER"},
-		{method: strings.Repeat("ohno", 9999), expected: "OTHER"},
-	}
-
-	for _, d := range tests {
-		actual := sanitizeMethod(d.method)
-		if actual != d.expected {
-			t.Errorf("Not same: expected %#v, but got %#v", d.expected, actual)
-		}
-	}
 }
